// dashboard/src/utils/hasura-api/generated/schemas/createActionArgs.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { ActionDefinition } from './actionDefinition';

export interface CreateActionArgs {
  /** Name of the action */
  name: string;
  definition: ActionDefinition;
  /** Custom comment. */
  comment?: string;
}