// dashboard/src/utils/hasura-api/generated/schemas/createActionStep.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { CreateActionArgs } from './createActionArgs';
import type { CreateActionStepType } from './createActionStepType';

/**
 * Create an action
 */
export interface CreateActionStep {
  type: CreateActionStepType;
  args: CreateActionArgs;
}